WIP: Send error 403 when SSO provider failed to authenticate the user
Redirect to login page on the client. Otherwise clients receive status 200 and can't detect the authentication error.

A GitHub webhook reported status 200 for requests even though it wasn't able to authenticate. I think Discourse's /session/sso_provider route should return an error 403 if the user isn't authenticated. The client could redirect to the login page if needed. Currently it returns 302 with a redirect to /login which is kinda wrong.